move old data file to deprecated
