+++ deprecated/depth_data.py
@@ -1,4 +1,4 @@
-from data.data import DaVinciDataSet, DaVinciDataModule
+from deprecated.data import DaVinciDataSet, DaVinciDataModule
 
 
 class DepthDaVinciDataSet(DaVinciDataSet):
